pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "backend project"
authors = [{ name = "Илья" }]
dependencies = [
    "fastapi",
    "sqlalchemy",
    "uvicorn[standard]",
    "dishka",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "pydantic[email]"
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "sqlalchemy-stubs",
    "pytest",
    "pre-commit"
]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I"]
ignore = ["E501", "F401"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
warn_unused_configs = true
warn_unused_ignores = true
